add excersise5

+++ src/main/java/excersise3/Divide.java
@@ -6,9 +6,9 @@ public class Divide {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("Input first number");
-        float firstNumber = input.nextFloat();
+        float firstNumber = input.nextInt();
         System.out.println("Input second number");
-        float secondNumber = input.nextFloat();
+        float secondNumber = input.nextInt();
         float divideResult = firstNumber / secondNumber;
         System.out.println(firstNumber + " / " + secondNumber + " = " + divideResult );
     }
